langchain-openai/requirements.txt
langchain-core
langchain-openai
langchain-community
python-dotenv
pypdf
faiss-cpu